requirements.txt
pyyaml
docopt
deepmerge